Extract getAllAsg
